requirements.txt
autopep8
PyYaml
torch
torchvision
azure-cli
azureml-core
azureml-dataprep
